Uptime: support JSON config

+++ src/detection/uptime/uptime_linux.c
@@ -7,6 +7,6 @@ const char* ffDetectUptime(uint64_t* result)
     struct sysinfo info;
     if(sysinfo(&info) != 0)
         return "sysinfo() failed";
-    *result = info.uptime;
+    *result = (uint64_t) info.uptime;
     return NULL;
 }
